shardmaster的server端超时去掉，改为默认的RPC超时，有两种BUG 1.因为之前在server 应用到日志之前就写了请求序列号SeqNum到history中去重，当网络波动导致请求丢失时,因为去重已写入，导致后续的请求被过滤无法完成。 2.Server端的请求超时，当一个请求已经被写入了到了raft，并且apply到了Server还没来得及返回时，Server 超时发生，返回了False，导致不断重试，但是被history再次过滤，客户端不返回，卡死。

+++ src/shardkv/common.go
@@ -35,7 +35,7 @@ type PutAppendArgs struct {
 	// You'll have to add definitions here.
 	// Field names must start with capital letters,
 	// otherwise RPC will break.
-	ClientId int
+	ClientId int64
 	SeqNum int
 }
 
@@ -47,7 +47,7 @@ type PutAppendReply struct {
 type GetArgs struct {
 	Key string
 	// You'll have to add definitions here.
-	ClientId int
+	ClientId int64
 	SeqNum int
 }
